Restructure project and extract starter packages
This commit restructures the project by extracting shared code into 'starter_toolkit' and 'starter_uikit' packages, moving and renaming files accordingly. It removes the old 'uikit' package, updates asset and localization paths, and introduces new features for environment and preferences management. The Flutter SDK is updated to 3.32.0, and obsolete files and assets are deleted. Test files are added for authentication and environment features.